Refactor codegen, test includes
