+++ example.vert
@@ -1,5 +1,8 @@
 #version 450
 
+// Relative include (resolved from current file)
+#include "tests/test.glsl"
+
 void main() {
-    gl_Position = vec4(0);
+    gl_Position = test(gl_Position);
 }
