+++ tests/test.glsl
@@ -1,0 +1,3 @@
+vec4 test(vec4 v) {
+    return v;
+}
